+++ src/onegov.user/onegov/user/__init__.py
@@ -8,11 +8,16 @@ _ = TranslationStringFactory('onegov.user')  # noqa
 from onegov.user.auth import Auth
 from onegov.user.collection import UserCollection
 from onegov.user.model import User
-from onegov.user.utils import is_valid_yubikey, yubikey_otp_to_serial
+from onegov.user.utils import (
+    is_valid_yubikey,
+    is_valid_yubikey_format,
+    yubikey_otp_to_serial
+)
 
 __all__ = [
     'Auth',
     'is_valid_yubikey',
+    'is_valid_yubikey_format',
     'User',
     'UserCollection',
     'yubikey_otp_to_serial'